2d dp
the first time I learned this problem, it was a part of my curriculum. dynamic programming was new to me at that time, let alone 2d dp approaches, so this problem was challenging for me. 
now that I have much better understanding of competitive coding in general, I would say it is very easy!